requirements.txt
Flask
black
pytest
pydantic
gunicorn
python-dotenv
Werkzeug
requests